09/12/23

+++ backend/src/main/java/br/com/docvet/docvet/service/TelefoneService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.docvet.docvet.domain.Telefone;
+import br.com.docvet.docvet.domain.error.NotFoundException;
 import br.com.docvet.docvet.repository.TelefoneRepository;
 
 @Service
@@ -15,7 +16,7 @@ public class TelefoneService {
     private TelefoneRepository repository;
 
     public Telefone getOne(Integer id){
-        return repository.findById(id).get();
+        return repository.findById(id).orElseThrow(()-> new NotFoundException("Teleofone não encontrado: id="+id));
     }
 
     public List<Telefone> getAll(){
@@ -27,14 +28,19 @@ public class TelefoneService {
     }
 
     public void update(Integer id, Telefone telefone){
-        Telefone telefoneExistente = repository.findById(id).get();
-
-        telefoneExistente.setNumero(telefone.getNumero());
-
-        repository.saveAndFlush(telefoneExistente);
+        if (!repository.existsById(telefone.getId())) {
+            throw new NotFoundException("Teleofone não encontrado: id="+telefone.getId());
+        }
+        // Telefone telefoneExistente = repository.findById(id).get();
+        // telefoneExistente.setNumero(telefone.getNumero());
+        repository.saveAndFlush(telefone);
     }
 
     public void delete(Integer id){
+        if (!repository.existsById(id)) {
+                        throw new NotFoundException("Teleofone não encontrado: id="+id);
+
+        }
         repository.deleteById(id);
     }
 }
